+++ pyodide_build/tests/test_buildpkg.py
@@ -263,3 +263,24 @@ def test_copy_sharedlib(tmp_path):
     deps = ("sharedlib-test.so", "sharedlib-test-dep.so", "sharedlib-test-dep2.so")
     for dep in deps:
         assert dep in dep_map
+
+
+def test_extract_tarballname():
+    url = "https://www.test.com/ball.tar.gz"
+    headers = [
+        {},
+        {"Content-Disposition": "inline"},
+        {"Content-Disposition": "attachment"},
+        {"Content-Disposition": 'attachment; filename="ball 2.tar.gz"'},
+        {"Content-Disposition": "attachment; filename*=UTF-8''ball%203.tar.gz"},
+    ]
+    tarballnames = [
+        "ball.tar.gz",
+        "ball.tar.gz",
+        "ball.tar.gz",
+        "ball 2.tar.gz",
+        "ball 3.tar.gz",
+    ]
+
+    for header, tarballname in zip(headers, tarballnames, strict=True):
+        assert buildpkg._extract_tarballname(url, header) == tarballname